# health-portal/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Patient Appointment and Vaccination Portal
  description: >
    RESTful API for managing patient profiles, immunization records, and appointment
    scheduling. Implements Richardson Maturity Model Level 3 (HATEOAS) with hypermedia
    links embedded in all resource representations.
  version: 1.0.0
servers:
  - url: https://health-portal-5699f82b1edc.herokuapp.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: List all patients
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of patients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequest'
      responses:
        '201':
          description: Patient created
          headers:
            Location:
              description: URI of the newly created patient
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /patients/{patientId}:
    parameters:
      - $ref: '#/components/parameters/PatientId'
    get:
      tags: [Patients]
      summary: Get a patient by ID
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Patients]
      summary: Update a patient profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequest'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Patients]
      summary: Delete a patient
      responses:
        '204':
          description: Patient deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Convenience: patient-scoped appointments (read-only)
...
  /patients/{patientId}/vaccinations:
    parameters:
      - $ref: '#/components/parameters/PatientId'
    get:
      tags: [Vaccinations]
      summary: List all vaccination records for a patient
      responses:
        '200':
          description: List of vaccination records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaccinationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Vaccinations]
      summary: Add a vaccination record for a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaccinationRequest'
      responses:
        '201':
          description: Vaccination record created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaccinationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /patients/{patientId}/vaccinations/{vaccinationId}:
    parameters:
      - $ref: '#/components/parameters/PatientId'
      - $ref: '#/components/parameters/VaccinationId'
    get:
      tags: [Vaccinations]
      summary: Get a specific vaccination record
      responses:
        '200':
          description: Vaccination record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaccinationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Vaccinations]
      summary: Update a vaccination record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaccinationRequest'
      responses:
        '200':
          description: Vaccination record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaccinationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Vaccinations]
      summary: Delete a vaccination record
      responses:
        '204':
          description: Vaccination record deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Appointments 
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Parameters 
  parameters:
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VaccinationId:
      name: vaccinationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # Reusable Responses 
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource already exists or constraint violated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # Schemas 
  schemas:
...
    Link:
      type: object
      properties:
        href:
          type: string
          format: uri
        rel:
          type: string
        method:
          type: string
          enum: [GET, POST, PUT, DELETE]
      required: [href, rel, method]
    Links:
      type: object
      description: HATEOAS hypermedia links
      additionalProperties:
        $ref: '#/components/schemas/Link'
      example:
        self:
          href: /api/v1/patients/abc-123
          rel: self
          method: GET
        update:
          href: /api/v1/patients/abc-123
          rel: update
          method: PUT
        delete:
          href: /api/v1/patients/abc-123
          rel: delete
          method: DELETE
        vaccinations:
          href: /api/v1/patients/abc-123/vaccinations
          rel: vaccinations
          method: GET
        schedule:
          href: /api/v1/appointments
          rel: schedule
          method: POST
    Address:
      type: object
      properties:
        street:
          type: string
          example: 123 Main St
        city:
          type: string
          example: Baltimore
        state:
          type: string
          example: MD
        zip:
          type: string
          example: "21201"
      required: [street, city, state, zip]
    VaccineType:
      type: string
      enum:
        - COVID_19
        - FLU
        - MMR
        - HEPATITIS_B
...
    PatientRequest:
      type: object
      properties:
        firstName:
          type: string
          example: Jane
        lastName:
          type: string
          example: Doe
        dateOfBirth:
          type: string
          format: date
          example: "1990-04-15"
        email:
          type: string
          format: email
          example: jane.doe@example.com
        phone:
          type: string
          example: "410-555-0100"
        address:
          $ref: '#/components/schemas/Address'
      required: [firstName, lastName, dateOfBirth, email, address]
    PatientResponse:
      allOf:
        - $ref: '#/components/schemas/PatientRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            _links:
              $ref: '#/components/schemas/Links'
          required: [id, _links]
    PatientListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PatientResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'
        _links:
          $ref: '#/components/schemas/Links'

    # Vaccination 
    VaccinationRequest:
      type: object
      properties:
        vaccineType:
          $ref: '#/components/schemas/VaccineType'
        dateAdministered:
          type: string
          format: date
          example: "2025-10-01"
        nextDueDate:
          type: string
          format: date
          example: "2026-10-01"
        provider:
          type: string
          example: City Health Clinic
        lotNumber:
          type: string
          example: A1B2C3
      required: [vaccineType, dateAdministered]
    VaccinationResponse:
      allOf:
        - $ref: '#/components/schemas/VaccinationRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            patientId:
              type: string
              format: uuid
            isDue:
              type: boolean
              description: True if nextDueDate is on or before today
            _links:
              $ref: '#/components/schemas/Links'
          required: [id, patientId, _links]
    VaccinationListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/VaccinationResponse'
        _links:
          $ref: '#/components/schemas/Links'

    # Appointment 
...
    PageMetadata:
      type: object
      properties:
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required: [timestamp, status, error, message, path]